refactor(mode): split ping transport from reachability policy
Why:
- IpcDaemonReachabilityProbe が ping送信と到達判定ポリシーを同時に持っており、変更理由が混在していたため。

What Changed:
- ping送信責務を IDaemonPingClient / IpcDaemonPingClient へ分離。
- IpcDaemonReachabilityProbe は事前チェック・タイムアウト・例外分類のポリシーに専念。
- Program の DI に ping client を追加して依存を配線。

Impact:
- mode 到達判定の凝集性が向上し、接続プロトコル変更時の影響範囲を局所化できる。
- 判定ロジックのテスト容易性が上がり、責務境界が明確になった。
