+++ src/Ucli/Execution/Mode/IpcDaemonReachabilityProbe.cs
@@ -1,6 +1,4 @@
 using System.Net.Sockets;
-using System.Text.Json;
-using MackySoft.Ucli.Contracts.Ipc;
 using MackySoft.Ucli.Foundation;
 using MackySoft.Ucli.Ipc;
 using MackySoft.Ucli.UnityProject;
@@ -10,26 +8,22 @@ namespace MackySoft.Ucli.Execution;
 /// <summary> Probes daemon reachability by attempting an IPC ping call. </summary>
 internal sealed class IpcDaemonReachabilityProbe : IDaemonReachabilityProbe
 {
-    private const string ProbeSessionToken = "mode-probe";
-
-    private const string ProbeClientVersion = "ucli-mode-probe";
-
     private static readonly TimeSpan ProbeTimeout = TimeSpan.FromMilliseconds(300);
 
     private readonly IIpcEndpointResolver endpointResolver;
 
-    private readonly IUnityIpcClient unityIpcClient;
+    private readonly IDaemonPingClient daemonPingClient;
 
     /// <summary> Initializes a new instance of the <see cref="IpcDaemonReachabilityProbe" /> class. </summary>
     /// <param name="endpointResolver"> The endpoint resolver dependency. </param>
-    /// <param name="unityIpcClient"> The Unity IPC client dependency. </param>
-    /// <exception cref="ArgumentNullException"> Thrown when <paramref name="endpointResolver" /> or <paramref name="unityIpcClient" /> is <see langword="null" />. </exception>
+    /// <param name="daemonPingClient"> The daemon ping client dependency. </param>
+    /// <exception cref="ArgumentNullException"> Thrown when <paramref name="endpointResolver" /> or <paramref name="daemonPingClient" /> is <see langword="null" />. </exception>
     public IpcDaemonReachabilityProbe (
         IIpcEndpointResolver endpointResolver,
-        IUnityIpcClient unityIpcClient)
+        IDaemonPingClient daemonPingClient)
     {
         this.endpointResolver = endpointResolver ?? throw new ArgumentNullException(nameof(endpointResolver));
-        this.unityIpcClient = unityIpcClient ?? throw new ArgumentNullException(nameof(unityIpcClient));
+        this.daemonPingClient = daemonPingClient ?? throw new ArgumentNullException(nameof(daemonPingClient));
     }
 
     /// <summary> Probes whether daemon for the specified project is reachable. </summary>
@@ -57,15 +51,12 @@ internal sealed class IpcDaemonReachabilityProbe : IDaemonReachabilityProbe
             return DaemonReachabilityProbeResult.NotRunning();
         }
 
-        var pingRequest = CreatePingRequest();
         using var timeoutCancellationTokenSource = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
         timeoutCancellationTokenSource.CancelAfter(ProbeTimeout);
         try
         {
-            await unityIpcClient.SendAsync(
-                    unityProject.UnityProjectRoot,
-                    unityProject.ProjectFingerprint,
-                    pingRequest,
+            await daemonPingClient.PingAsync(
+                    unityProject,
                     timeoutCancellationTokenSource.Token)
                 .ConfigureAwait(false);
             return DaemonReachabilityProbeResult.Running();
@@ -88,21 +79,6 @@ internal sealed class IpcDaemonReachabilityProbe : IDaemonReachabilityProbe
             return DaemonReachabilityProbeResult.Failure(
                 ExecutionError.InternalError($"Failed to probe daemon reachability. {exception.Message}"));
         }
-    }
-
-    /// <summary> Creates one IPC ping request used for reachability probing. </summary>
-    /// <returns> The ping request envelope. </returns>
-    private static IpcRequest CreatePingRequest ()
-    {
-        var payload = JsonSerializer.SerializeToElement(
-            new IpcPingRequest(ProbeClientVersion),
-            IpcJsonSerializerOptions.Default);
-        return new IpcRequest(
-            ProtocolVersion: IpcProtocol.CurrentVersion,
-            RequestId: $"mode-probe-{Guid.NewGuid():N}",
-            SessionToken: ProbeSessionToken,
-            Method: IpcMethodNames.Ping,
-            Payload: payload);
     }
 
     /// <summary> Determines whether an exception indicates endpoint reachability failure. </summary>
