+++ src/Ucli/Program.cs
@@ -50,6 +50,7 @@ internal static class Program
                 services.AddSingleton<IInitService, InitService>();
                 services.AddSingleton<IIpcEndpointResolver, IpcEndpointResolver>();
                 services.AddSingleton<IUnityIpcClient, UnityIpcClient>();
+                services.AddSingleton<IDaemonPingClient, IpcDaemonPingClient>();
                 services.AddSingleton<IDaemonReachabilityProbe, IpcDaemonReachabilityProbe>();
                 services.AddSingleton<IUnityExecutionModeDecisionService, UnityExecutionModeDecisionService>();
                 services.AddSingleton<IOperationCatalogProvider, InMemoryOperationCatalogProvider>();
